Initial draft of the libxkbregistry API
This is just a header file and the basic boilerplate to build the doxygen
(taken from libxbkcommon).

The API itself should hold few surprises, it's basically a C struct mapping to
the current XML data combined with the context initialization that
libxkbcommon uses.

The 'rxkb' prefix was chosen because 'xkb_registry' is too long and 'xkbr' is
not visually distinct enough from libxkbcommon's 'xkb' prefix.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>